requirements.txt
# Core dependencies
netmiko>=4.3.0
paramiko>=3.4.0
jinja2>=3.1.2
python-dotenv>=1.0.0

# Optional but recommended
rich>=13.7.0
colorama>=0.4.6

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
flake8>=6.1.0
black>=23.12.0
pylint>=3.0.3

# Security
bandit>=1.7.5

# Utilities
pyyaml>=6.0.1
